Make AdvancedFilter separate component
* Also add missing types
