+++ app/src/views/private/components/advanced-filter/get-available-operators-for-type.ts
