+++ app/src/views/private/components/advanced-filter/index.ts
@@ -1,0 +1,4 @@
+import AdvancedFilter from './advanced-filter.vue';
+
+export { AdvancedFilter };
+export default AdvancedFilter;
