+++ app/src/views/private/components/advanced-filter/types.ts
@@ -1,0 +1,12 @@
+import { Type } from '@directus/shared/types';
+
+export type FieldTree = {
+	field: string;
+	name: string;
+	children?: FieldTree[];
+};
+
+export type OperatorType = {
+	type: Type;
+	operators: string[];
+};
