--- app/src/views/private/components/filter-sidebar-detail/types.ts
@@ -1,5 +1,0 @@
-export type FieldTree = {
-	field: string;
-	name: string;
-	children?: FieldTree[];
-};
